+++ src/app/cli/common_args.go
@@ -6,12 +6,8 @@ import (
 	"time"
 )
 
-type MultipleFilesArgs struct {
-	File []string `arg required type:"existingfile" name:"file" help:".klg source file(s)"`
-}
-
-type SingleFileArgs struct {
-	File string `arg required type:"existingfile" name:"file" help:".klg source file"`
+type InputFilesArgs struct {
+	File []string `arg optional type:"existingfile" name:"file" help:".klg source file(s) (falls back to bookmark if not specified)"`
 }
 
 type DiffArg struct {